+++ src/main/java/micycle/spacefillingcurves/SierpinskiFourSteps.java
@@ -4,7 +4,7 @@ import org.locationtech.jts.geom.Coordinate;
 
 public class SierpinskiFourSteps extends SpaceFillingCurve {
 
-	public SierpinskiFourSteps(int width, int height) {
+	public SierpinskiFourSteps(double width, double height) {
 		super(width, height);
 	}
